More debug console
Added drawtile command to the debug console, which is what started this whole thing! I found lots of issues with this implementation of the console while working on adding this command. I'm not sure it's worth dwelling on right now - it's just really frustrating how few things you can do with strings in GML.